refactor: encapsulate all provider interactions inside SandboxResource
Move lifecycle state machine and exec into SandboxResource so that
every sandbox operation requires an Authenticator, preventing
cross-workspace usage. The provider is resolved internally and never
exposed to consumers.

- Add SandboxResource.ensureActive() (absorbs lifecycle.ts)
- Add SandboxResource.exec() with workspace ownership assertion
- Catch CommandExitError inside E2B provider (no E2B leak in tools)
- Delete lifecycle.ts
- Simplify tool handler to only use SandboxResource
